Generic construct for blueprint builder
