+++ lib/stacks/index.ts
@@ -1,1 +1,2 @@
 export * from './eks-blueprint-stack';
+export * from './eks-blueprint-construct';
